chagned to async-await, removed import, changed folder
